workers/caption/requirements.txt
runpod
torch>=2.2.0
transformers>=4.53.0
accelerate>=0.30.0
safetensors>=0.4.0
Pillow>=10.0.0
qwen-vl-utils
sentencepiece
